Share some binary syntax code between JIT back-ends
Break out code for binary construction and matching that do not
depend on the architecture (x86_64 or AArch64).
