+++ erts/emulator/Makefile.in
@@ -1056,6 +1056,7 @@ JIT_OBJS = \
     $(OBJDIR)/asm_load.o \
     $(OBJDIR)/beam_asm_global.o \
     $(OBJDIR)/beam_asm_module.o \
+    $(OBJDIR)/beam_jit_bs.o \
     $(OBJDIR)/beam_jit_common.o \
     $(OBJDIR)/beam_jit_main.o \
     $(OBJDIR)/beam_jit_metadata.o \
